+++ android/src/main/kotlin/com/github/florent37/assets_audio_player/AssetsAudioPlayerPlugin.kt
@@ -135,7 +135,7 @@ class AssetsAudioPlayer(
 
     fun unregister() {
         stopWhenCall?.stop()
-        notificationManager.hideNotification(definitively = true)
+        notificationManager.hideNotificationService(definitively = true)
         stopWhenCall?.unregister(stopWhenCallListener)
         players.values.forEach {
             it.stop()
@@ -290,6 +290,23 @@ class AssetsAudioPlayer(
                         return
                     }
                     getOrCreatePlayer(id).setPlaySpeed(speed)
+                    result.success(null)
+                } ?: run {
+                    result.error("WRONG_FORMAT", "The specified argument must be an Map<*, Any>.", null)
+                    return
+                }
+            }
+            "showNotification" -> {
+                (call.arguments as? Map<*, *>)?.let { args ->
+                    val id = args["id"] as? String ?: run {
+                        result.error("WRONG_FORMAT", "The specified argument (id) must be an String.", null)
+                        return
+                    }
+                    val show = args["show"] as? Boolean ?: run {
+                        result.error("WRONG_FORMAT", "The specified argument (show) must be an Boolean.", null)
+                        return
+                    }
+                    getOrCreatePlayer(id).showNotification(show)
                     result.success(null)
                 } ?: run {
                     result.error("WRONG_FORMAT", "The specified argument must be an Map<*, Any>.", null)