+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.1+6
+
+* Added `.showNotification = true/false` to hide dynamically displayed notification
+
+## 2.0.1+4
+
+* Added `isBuffering` to `RealtimePlayingInfos`
+
 ## 2.0.1+3
 
 * Added `AssetsAudioPlayerGroup` (beta)